Remove linkme dependency need.
