+++ crates/zng-view-prebuilt/Cargo.toml
@@ -21,7 +21,6 @@ embedded = []
 zng-env = { path = "../zng-env", version = "0.1.1" }
 libloading = "0.8"
 dunce = "1.0"
-linkme = "0.3"
 
 [build-dependencies]
 hashers = "1.0"
