+++ examples/extend-view/Cargo.toml
@@ -12,4 +12,3 @@ zng-app = { path = "../../crates/zng-app" }
 tracing = "0.1"
 serde = { version = "1.0", features = ["derive"] }
 raw-window-handle = "0.6"
-linkme = "0.3"
